// src/modules/garmentsProduction/reports/cuttting/date-wise-fabric-requisition-receive/date-wise-fabric-requisition-receive-type.ts
export interface IDateWiseFabricRequisitionReceive {
  BUYER_ID: number;
  BUYER_NAME: string;
  PO_ID: number;
  PO_NO: string;
  STYLE_NO: string;
  COLOR_ID: number;
  COLOR_NAME: string;
  FABRIC_ID: number;
  FABRIC_NAME: string;
  UNIT: string;
  FABRIC_SOURCE: string;
  TOTAL_REQUIRED: number;
  TOTAL_RECEIVE: number;
  READY_QTY: number;
  DAY_REQUISITION: number;
  DAY_RECEIVE: number;
}